refactor(utils): update typing imports
